Update handlebars to version 3.0.0.
