+++ framework/core/app.js
@@ -43,7 +43,7 @@ app = (function() {
             bootstrap: 'core/libraries/bootstrap-3.3.2',
             'bootstrap.switch': 'core/libraries/bootstrap.switch-3.3.2',
             fastclick: 'core/libraries/fastclick-1.0.6',
-            handlebars: 'core/libraries/handlebars-2.0.0',
+            handlebars: 'core/libraries/handlebars-3.0.0',
             jasmine: 'core/libraries/jasmine-2.1.3',
             'jasmine.html': 'core/libraries/jasmine.html-2.1.3',
             jquery: 'core/libraries/jquery-2.1.3',
